add shopinvader_sale_state and migrate shopinvader_delivery_state
