+++ .pre-commit-config.yaml
@@ -23,7 +23,6 @@ exclude: |
   ^shopinvader_customer_validate/|
   ^shopinvader_delivery_carrier/|
   ^shopinvader_delivery_instruction/|
-  ^shopinvader_delivery_state/|
   ^shopinvader_demo_app/|
   ^shopinvader_easy_binding/|
   ^shopinvader_elasticsearch/|
