+++ setup/shopinvader_delivery_state/odoo/addons/shopinvader_delivery_state
@@ -1,0 +1,1 @@
+../../../../shopinvader_delivery_state
