+++ setup/shopinvader_sale_state/odoo/addons/shopinvader_sale_state
@@ -1,0 +1,1 @@
+../../../../shopinvader_sale_state
