+++ shopinvader_restapi/README.rst
@@ -7,7 +7,7 @@ Shopinvader
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:9694f0bb4c0070ebfb5396b20a53e51dd0061395c798df80e84732afb4f03280
+   !! source digest: sha256:5983a00bf1e9d8b6ea50348bedd76002794098067f07d54fd3a185e2059a445f
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Alpha-red.png
