+++ shopinvader_restapi/models/sale.py
@@ -5,7 +5,7 @@
 
 import logging
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 _logger = logging.getLogger(__name__)
 
@@ -23,38 +23,6 @@ class SaleOrder(models.Model):
         string="Done Cart Step",
         readonly=True,
     )
-    # TODO move this in an extra OCA module
-    shopinvader_state = fields.Selection(
-        [
-            ("cancel", "Cancel"),
-            ("pending", "Pending"),
-            ("processing", "Processing"),
-            ("shipped", "Shipped"),
-        ],
-        compute="_compute_shopinvader_state",
-        store=True,
-    )
-
-    def _get_shopinvader_state(self):
-        self.ensure_one()
-        if self.state == "cancel":
-            return "cancel"
-        elif self.state == "done":
-            return "shipped"
-        elif self.state == "draft":
-            return "pending"
-        else:
-            return "processing"
-
-    def _compute_shopinvader_state_depends(self):
-        return ("state",)
-
-    @api.depends(lambda self: self._compute_shopinvader_state_depends())
-    def _compute_shopinvader_state(self):
-        # simple way to have more human friendly name for
-        # the sale order on the website
-        for record in self:
-            record.shopinvader_state = record._get_shopinvader_state()
 
     def _prepare_invoice(self):
         res = super(SaleOrder, self)._prepare_invoice()
