+++ shopinvader_sale_state/README.rst
@@ -1,32 +1,34 @@
-==========================
-Shopinvader delivery state
-==========================
+======================
+Shopinvader Sale State
+======================
 
 .. 
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:cb29f1c7d2638ae4b19ed12f0b0053314965e4efc250713ef00aceac8ccfbc21
+   !! source digest: sha256:558b9628b4e0fecdf10fcf63d5e68277f0fd3b312cf4525621c2ce5a1ee6324c
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+.. |badge1| image:: https://img.shields.io/badge/maturity-Alpha-red.png
     :target: https://odoo-community.org/page/development-status
-    :alt: Beta
+    :alt: Alpha
 .. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-shopinvader%2Fodoo--shopinvader-lightgray.png?logo=github
-    :target: https://github.com/shopinvader/odoo-shopinvader/tree/16.0/shopinvader_delivery_state
+    :target: https://github.com/shopinvader/odoo-shopinvader/tree/16.0/shopinvader_sale_state
     :alt: shopinvader/odoo-shopinvader
 
 |badge1| |badge2| |badge3|
 
-Integrate `sale_delivery_state` with Shopinvader.
+This addon adds the shopinvader state on sale order.
+This state will be the state on the website and can be extended by extra module
 
-When `sale_delivery_state` is installed this module will be automatically installed.
-
-This way the shipping state of shop orders will be computed properly.
+.. IMPORTANT::
+   This is an alpha version, the data model and design can change at any time without warning.
+   Only for development or testing purpose, do not use in production.
+   `More details on development status <https://odoo-community.org/page/development-status>`_
 
 **Table of contents**
 
@@ -39,7 +41,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/shopinvader/odoo-shopinvader/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/shopinvader/odoo-shopinvader/issues/new?body=module:%20shopinvader_delivery_state%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/shopinvader/odoo-shopinvader/issues/new?body=module:%20shopinvader_sale_state%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -49,24 +51,16 @@ Credits
 Authors
 ~~~~~~~
 
-* Camptocamp SA
+* Akretion
 
 Contributors
 ~~~~~~~~~~~~
 
-* Simone Orsi <simone.orsi@camptocamp.com>
-
-Other credits
-~~~~~~~~~~~~~
-
-The development of this module has been financially supported by:
-
-* Camptocamp
-* Cosanum
+* Sébastien BEAU <sebastien.beau@akretion.com>
 
 Maintainers
 ~~~~~~~~~~~
 
-This module is part of the `shopinvader/odoo-shopinvader <https://github.com/shopinvader/odoo-shopinvader/tree/16.0/shopinvader_delivery_state>`_ project on GitHub.
+This module is part of the `shopinvader/odoo-shopinvader <https://github.com/shopinvader/odoo-shopinvader/tree/16.0/shopinvader_sale_state>`_ project on GitHub.
 
 You are welcome to contribute.
