+++ shopinvader_sale_state/__manifest__.py
@@ -2,23 +2,25 @@
 # @author Sébastien BEAU <sebastien.beau@akretion.com>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
+
 {
-    "name": "Shopinvader Schema Sale",
-    "summary": "Add schema sale",
-    "version": "16.0.1.1.2",
+    "name": "Shopinvader Sale State",
+    "summary": "Basic module to implement state for sale order",
+    "version": "16.0.1.0.0",
     "development_status": "Alpha",
+    "category": "Uncategorized",
     "website": "https://github.com/shopinvader/odoo-shopinvader",
     "author": " Akretion",
     "license": "AGPL-3",
-    "depends": [
-        "sale_cart",
-        "sale_discount_display_amount",
-        "shopinvader_sale_state",
-        "shopinvader_schema_address",
-    ],
     "external_dependencies": {
-        "python": ["extendable_pydantic>=1.2.0", "pydantic>=2.0.0"]
+        "python": [],
+        "bin": [],
     },
-    "data": [],
+    "depends": [
+        "sale",
+    ],
+    "data": [
+        "views/sale_view.xml",
+    ],
     "demo": [],
 }
