+++ shopinvader_sale_state/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Sébastien BEAU <sebastien.beau@akretion.com>
