+++ shopinvader_sale_state/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This addon adds the shopinvader state on sale order.
+This state will be the state on the website and can be extended by extra module
