+++ shopinvader_sale_state/views/sale_view.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+
+    <record id="view_order_tree" model="ir.ui.view">
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale.view_order_tree" />
+        <field name="arch" type="xml">
+            <field name="invoice_status" position="after">
+                <field
+                    name="shopinvader_state"
+                    optional="hide"
+                    decoration-info="shopinvader_state in ['pending', 'processing']"
+                    decoration-success="shopinvader_state == 'shipped'"
+                    widget="badge"
+                />
+            </field>
+        </field>
+    </record>
+
+    <record id="view_quotation_tree" model="ir.ui.view">
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale.view_quotation_tree" />
+        <field name="arch" type="xml">
+            <field name="state" position="after">
+                <field
+                    name="shopinvader_state"
+                    optional="hide"
+                    decoration-info="shopinvader_state in ['pending', 'processing']"
+                    decoration-success="shopinvader_state == 'shipped'"
+                    widget="badge"
+                />
+            </field>
+        </field>
+    </record>
+
+</odoo>
