+++ shopinvader_schema_sale/README.rst
@@ -7,7 +7,7 @@ Shopinvader Schema Sale
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:97ce698e11ec268908c328fcdc1ed151977e0f218c182836b51ba8189357a3c5
+   !! source digest: sha256:5fdaa088b0b5c5b3d4f2b8bfc7331d6c7ed4d5ab7d649b54af0d44022f13e5cb
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Alpha-red.png
